Move obj.py to _internal
