+++ certbot-apache/certbot_apache/tests/centos_test.py
@@ -7,7 +7,7 @@ from certbot import errors
 from certbot.compat import filesystem
 from certbot.compat import os
 
-from certbot_apache import obj
+from certbot_apache._internal import obj
 from certbot_apache._internal import override_centos
 from certbot_apache.tests import util
