+++ certbot-apache/certbot_apache/tests/debian_test.py
@@ -8,7 +8,7 @@ from certbot import errors
 from certbot.compat import os
 
 from certbot_apache._internal import apache_util
-from certbot_apache import obj
+from certbot_apache._internal import obj
 from certbot_apache.tests import util
