+++ certbot-apache/certbot_apache/tests/display_ops_test.py
@@ -9,7 +9,7 @@ from certbot.display import util as display_util
 
 from certbot.tests import util as certbot_util
 
-from certbot_apache import obj
+from certbot_apache._internal import obj
 
 from certbot_apache._internal.display_ops import select_vhost_multiple
 from certbot_apache.tests import util
