+++ certbot-apache/docs/api/obj.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_apache.obj`
+:mod:`certbot_apache._internal.obj`
 -----------------------------
 
-.. automodule:: certbot_apache.obj
+.. automodule:: certbot_apache._internal.obj
    :members:
